Configure Bugsnag in client-side code
If the Bugsnag API key is exposed to the client, then configure
client-side reporting. By default (as with server-side code), Bugsnag
will report uncaught exceptions and unhandled rejections.

Additionally, the Bugsnag React plugin includes an error boundary that
reports errors. It functions very similarly to the react-error-boundary
library we were previously using.

However, it does require that Bugsnag be configured and enabled. Given
that we're treating Bugsnag as an optional dependency, we should have a
functioning error boundary regardless of whether or not Bugsnag is
configured. Therefore, use Bugsnag's error boundary if Bugsnag is
configured; otherwise continue using react-error-boundary.