Fix issue preventing updating Hunt/Puzzle when url absent
Both Hunt and Puzzle contain a field which is optional, but if present,
should be a valid URL.  For both of these objects, when we would perform
updates on the object, we `$set` the provided keys and `$unset` the
others.  Our custom validator was failing since `this.isSet` was true,
and empty-string does not parse as a valid URL.  When `this.operator` is
`$unset`, though, we don't care about the value -- unsetting the field
is always valid.