Introduce `TypedPublication`
`TypedPublication` is a lot like `TypedMethod`, but, as you might
expect, for publications.

Unlike methods, publications don't need to have a name. If you publish
with a name of `null`, the publication is automatically pushed to
clients. This is supported through a `DefaultTypedPublication`, which
requires that the publication take no arguments.

We could theoretically continue to use `Meteor.publish` directly for
those, but `TypedPublication` also introduces error handling, which
seemed useful for all publications, not just ones that could be
explicitly subscribed. As it turns out, Meteor by default more or less
completely swallows any exceptions thrown by a publish function.
(Specifically, it cancels the subscription and sends the exception to
the client in a "nosub" DDP message, which the client completely
ignores)