Fix validation for server-side method invocations
When invoked from the client, method arguments are encoded with EJSON
before being sent over DDP. This has the specific side-effect of
stripping any key-value pairs where the value is undefined.

This doesn't happen when methods are invoked on the server. That
interacts poorly with `Match.Optional`, which permits keys to be absent
but does _not_ permit them to be `undefined`.

To make the behavior consistent, round-trip arguments through EJSON with
server-side invocations.

Fixes #1297.