Fix server-side source code capture for Bugsnag (Take 2)
If you supply a projectRoot, Bugsnag will find source code in that
directory and include it as part of the error report. This doesn't work
out of the box in production because we don't have the non-compiled
source anymore. However, we do have it in the form of sourcemaps (which
contain the full uncompiled source code). So use the source maps to
materialize a temporary directory to use as a projectRoot for Bugsnag.

(We can't get Bugsnag to use the sourcemaps because Meteor installs
source-map-support, which rewrites the stack traces before Bugsnag gets
a chance to see them, so Bugsnag can't map the post-rewrite stack frame
onto the source map, which, since it refers to the post-compilation
filename, doesn't match)

Unlike the first attempt in 7107fb74, this drops the source-map
repository, which uses wasm for faster parsing of mappings data. That
isn't compatible with Meteor's use of fibers, but we don't need it, so
it doesn't matter.