Add a content field to ChatMessage schema
I tried making the schema reflect everything, but our simpl-schema
integration had trouble with the fields, and it's a standard JSON
document anyway, so it's not much trouble to stringify and parse
wherever we want to use it.

I did write types for the actual structured messages, so we can cast
once and have checking on intermediate structure at least.