Remove Ansible from client-side code
In spite of its name (the original goal was sending FTL messages from
client to server), Ansible has always been more heavily used on server
code than client code, so make it official. Replace the handful of
places we used Ansible on the client with regular logs (supplemented by
Bugsnag reporting if appropriate) or move the logs to server-side
methods.