Automatically re-join calls after code-push-triggered reload
If Meteor triggers a page reload due to out-of-date code, stash the
current state of whatever call we're in in sessionStorage, and use that
to automatically re-join the call afterwards. We do some shenanigans to
ensure that this only happens when Meteor triggers a page reload (as
opposed to the user refreshing the page on their own).