Block code reloads when it would destroy state
When a user is in a call or in the middle of composing a chat message,
that's a particularly unfortunate time to reload the page due to a new
code push.

So, introduce a new hook that can hold code updates. So long as a
non-undefined string is passed to `useBlockUpdate`, any new code push
will cause a notification to pop up, but the page won't refresh.