Prevent rerendering entire chat history
Whenever the data returned by a useTracker changes, the entire data set
is replaced. This means that even if (e.g.) the only change was adding a
new element to an array, all other elements of that array will return
new objects, which won't compare as equal to the old objects (even if
they have the same contents).

This means in practice that any time (e.g.) a new chat message is
created, the entire chat history has to be re-rendered.

By using the useFind hook instead, we can significantly decrease the
amount of work required. In my testing, this reduced render time by
factor of roughly 10.