Use Mediasoup to heartbeat between servers
Instead of heartbeating through MongoDB writes, we can use Mediasoup
data producers and consumers to heartbeat between servers directly.

Each server creates a dedicated router for handling heartbeats; a
DirectTransport, which allows the Meteor process to send data packets
back and forth to the Mediasoup worker without a full WebRTC
negotiation; and a DataProducer over that DirectTransport, which emits a
heartbeat packet every 100ms (+/- jitter).

Then, whenever a server notices another server in the list of servers,
it establishes a PipeTransport connection to that other server and pipes
the heartbeat producer over that transport. It creates a
MonitorConnectRequest record with the details of its half of the
transport. Finally, it sets a watchdog - if it does not receive a
heartbeat from the other server, it will mark that server as dead (and
trigger the garbage collection cleanup hooks). The watchdog has an extra
grace period before receiving the first heartbeat to allow for any
potential delays in setting up the connection.

When a server observes a new MonitorConnectRequest record, it creates
and connects its half of the PipeTransport and writes out a
MonitorConnectAck so that the initiating server can connect, too. Then
it resets the watchdog every time it receives a heartbeat packet.

Two notes about how Mediasoup works in this configuration:

First, as a reminder, Mediasoup's use of "producer" and "consumer" is
from the perspective of a client, which makes it especially confusing
here, because the sender of the heartbeats creates a "consumer" on the
pipe transport, and the receiver creates a "producer".

Second, since PipeTransports don't use ICE to negotiation a connection,
so you have to provide a single IP address, rather than a list of
potential IP addresses, so we just grab the first one. In production,
since we only use publicly-routable addresses for ICE, this should be
fine.

Since this should allow us to detect dying servers at an acceptable pace
for recovering WebRTC calls, we can also significantly relax the
MongoDB-based heartbeats back to their previous frequency (heartbeats
every 15 seconds with a 120 second expiration window).

Fixes #1205.