Cache Discord permission grants to limit rate limiting
Discord has a fairly low rate limit, so blindly attempting to add anyone
in the hunt to a role is a quick road to aggressive rate limiting.
Instead, keep track of who we've previously granted a role to (similar
to what we do for Google Drive permissions) and avoid trying to re-grant
them their role.

However, while we populate the cache any time we add a user to a role,
we only check the cache for the syncHuntDiscordRole method (and for the
updateHunt method). In any other case, we're acting on behalf of a
specific user, and we should take action for them even if we might have
done it already (to ensure they don't get stuck in a bad state
accidentally).

And while it's not my preferred solution, arguably this does fix #1296.