Don't show guesses in guess queue for hunts you don't operate
The previous behavior was that NotificationCenter would show any pending
guess that was visible to the client.  This is slightly too broad: when
viewing a puzzle, all guesses for that puzzle are fetched, regardless of
state, so if you were an operator in hunt A you'd see the guesses for a
puzzle you had open in hunt B, even though you can't action them.

Instead, we should have NotificationCenter filter guesses down to those
that the user can action, which is the set of those from hunts the user
is an operator for.

Fixes #1353.