Fix findOne{,Allowing}DeletedAsync methods
Because of how Meteor currently implements their async methods, they
call into our implementation of `findOneAsync`, which injects the
deleted=false term to all queries, including those where we explicitly
want deleted (or both deleted and non-deleted) objects.

The easiest way to work around that right now is to implement our own
versions of the async methods in the same style as Meteor's - simply
wrapping the synchronous versions.

This will eventually break as Meteor continues to migrate towards async
implementations by default, so add a test that will fail if and when
that happens.