Add additional client-side Bugsnag reporting
This catches any Meteor method exception and reports it to Bugsnag. This
will be duplicative with our server-side reporting, but we generally
have significantly better contextual information on the client, which
should make debugging easier.

Additionally, add breadcrumbs on method calls so we can better tell what
was happening immediately prior to an error.