Hide `Guess`es that reference deleted `Puzzle`s
We will simply pretend that they do not exist, and avoid showing them in
the UI.

We also prevent taking actions on such Guesses -- we don't want to wind
up running solved/unsolved hooks, and it's somewhat nonsensical to try
to transition the guess state of a Guess that isn't displayed anywhere
anyway.

Fixes #1371.