Rework lifecycle of JoinPublisher
Previously, JoinPublisher would call subscription.ready(), but the
caller would have to wire up the shutdown method. This is sort of silly,
since JoinPublisher is perfectly capable of wiring up its own shutdown.

Also, make it a function instead of a class. We don't actually need to
capture the returned object - it previously was mostly just a handle to
the shutdown function, and since it wires up its own shutdown, we don't
need that anymore. And eslint gets mad at you if you try and instantiate
a class purely for its side effects.

Additionally, rework pendingGuessesForSelf to take advantage of
publishJoinedQuery + PublicationMerger.