Replace Base.publish with TypedPublications
There were only about 20 remaining places where we used our old
"mongo.*" publications from Base, so convert all of those to
TypedPublications.

I have a long-term goal of using the various publication infrastructure
we've built up (TypedPublication, publishJoinedQuery, etc.) to better
optimize how much data we're streaming to clients. Therefore, in as many
cases as I could, I tried to create publications specialized to the
pages loading them, including (e.g.) replacing bulk fetches with
appropriate joins.

Most of these are comparatively straightforward; the most complicated
was on PuzzlePage, where we previously recombined multiple subscriptions
into multiple loading states without any being a clean subset of the
other. However, the only reason for crossover was to determine if the
puzzle had been deleted (so that we could disable chat and voice).
Instead, I restructured the code to _assume_ that the puzzle was deleted
until we confirmed that it wasn't, which removed the explicit dependency
on loading state.