Add a utility for merging overlapping publications
In a few cases, we want to be able to have multiple publications with
potentially overlapping datasets. If multiple separate subscriptions
publish the same dataset over the same connection, Meteor will merge
those, but there's no way to access that functionality if you want to
have multiple processes within a single publication touching the same
dataset.

To mediate this, introduce PublicationMerger, with an implementation
heavily borrowed from Meteor's, which can spin off "sub-subscriptions",
each of which has independent state.

This also supports removing a sub-subscription, which is half of the
functionality we need today from SwappableCursorPublisher.