Fix error in eslint check
When I first wrote this, I didn't understand why there were types which
weren't either classes or interfaces, and yet I was able to get their
base types. As a result, I made some (incorrect) assumptions that any
ObjectType could be passed to getBaseTypes.

I now know that the types I previously had problems with were
TypeReferences, which are used with generics. Instead of trying to call
getBaseTypes on any ObjectType (which doesn't work), unwrap
TypeReferences that we find.

(Also enable sourcemaps for the compiled rule while we're in the
neighborhood.)