Introduce infrastructure for Zod-based schemas
This is a new system for defining database schemas which exist both at
the type layer and the runtime layer. Instead of using io-ts to define
the type layer and compiling that into SimpleSchema (which enforces the
runtime layer), we instead use Zod to define schemas and compile them
into JSON Schema, which we can then attach directly to the Mongo
database to enforce the runtime behavior.

Zod represents an improvement for our use cases over io-ts in a few
ways:
- Its syntax for expressing optional types is much more concise (no more
  `t.union([t.string, t.undefined])`)
- It allows expressing additional non-purely-type constraints, such as
  regex matches or lengths (which we previously expressed in Overrides)
- It is somewhat more pleasant to metaprogram over at runtime (e.g. to
  generate JSON Schema)

Meanwhile, JSON Schema is capable of expressing everything we currently
want to support in our schema language (unlike SimpleSchema which can't
support, e.g., tagged unions) and is relatively straightforward to
recursively generate from the Zod schema object (unlike SimpleSchema,
which makes it fairly challenging to express things like nested object
types).

The biggest downside of this approach is that while SimpleSchema comes
with a native understanding of Mongo operations, giving it built-in
support for auto-populating fields (e.g. `createdAt` timestamps) and
standardized transformations on values which worked across different
Mongo operators (e.g. our `answerify` function for uppercasing answers),
JSON Schema supports no runtime transformation, only validation, and Zod
has no understanding of Mongo operations to apply transformations
intelligently.

To work around this, we introduce logic which can take a Zod schema and
manipulate it such that it's possible to use it to transform Mongo
operations. There are three main moving parts to make this work:
1. We have to relax certain constraints on the schema (e.g. a required
   field may not be present in an update operation). This potentially
   means that our runtime validation will allow through an invalid
   operation, but since our primary goal is transformation and not
   validation, that's acceptable. (Ultimately, the JSON schema
   enforcement at Mongo itself ensures validity.) We can do this in
   advance in the `relaxSchema` function.
2. We have to account for Mongo's dot-separated addressing syntax, which
   can put nested fields at the top level of an update operation. To
   account for this, we group the fields of the update operation by
   their top-level key (or lack of key, for non-dot-separated fields).
   For each group, we generate a new schema which represents just that
   sub-object (or array element, etc.), and then apply that schema to
   the group, recursing as necessary. Because we can't enumerate all
   possible ways that nested fields can be addressed (e.g. arrays can be
   indexed with a number), we have to do this at runtime, although that
   should only have cost if we're actually updating a nested field. This
   is handled by `parseMongoOperationAsync` (with assistance from
   `getSchemaForField`)
3. We need to ensure that auto-populated fields are populated at the
   appropriate time (be that insert, upsert, or update). For this, we
   use Meteor's EnvironmentVariables (which are basically variables with
   dynamic scope) to flag what state we're currently in, and then use
   Zod's transform functions to populate fields at the appropriate
   times. We set the insert flag on inserts, the upsert flag when
   generating `$setOnInsert` in an update, and the update flag for the
   `$set` operation in an update. (This is handled sort of all over the
   place, but primarily through cooperation between
   `parseMongoModifierAsync` and `imports/lib/schemas/customTypes.ts`.)

On top of the environment variables, an auto-populated field requires us
to make a distinction between the type that is returned from the
schema's processing and the type that is guaranteed to be present in the
database. Under normal circumstances, the type that is returned from the
schema's processing would be Zod's "output" type (`z.output` or
`z.infer`). However, that type isn't especially useful to us, so we
repurpose the output type to instead represent the type that is
guaranteed to be present in the database. This requires us to, frankly,
lie to Zod and the type system by using transform functions to claim
that certain fields are not nullable, when in fact they are. This
happens in `imports/lib/schemas/customTypes.ts`.

However, we retain the characteristic of Zod that the input type
(`z.input`) represents what the user is obligated to provide, at least
for insertion of a new document. (As outlined above, user obligations
around updates are a bit more...fuzzy.)

(Additionally, because Zod's transformations and other "effects" suffer
from runtime type erasure, we need an additional annotation to generate
the correct JSON schema.)

All of this functionality is wrapped up in a new `Model` class, intended
to replace both the old `Base` class and direct usage of
`Mongo.Collection`. Unlike `Base`, it wraps a Mongo collection, rather
than extending it, giving us more control over the interfaces we choose
to support. For the specific `Base` functionality of supporting
soft-deletion, there's a separate `SoftDeletedModel` class.

Rather than requiring that schemas provided to them hew to a specific
interface, both `Model` and `SoftDeletedModel` add fields (`_id`,
`deleted`) that they require to be present on their schemas. This means
that the Zod schema object can not on its own be used to derive a
model's type, so we instead export a new `ModelType` helper, replacing
the pattern of exporting `t.TypeOf<typeof MySchema>` in the schema file.

`Model` additionally includes a few additional affordances. First of
all, it requires that all string types be non-empty (either via length
constraint or via regex). This is an alternative to SimpleSchema's
default behavior of replacing empty strings with undefined and has the
benefit of being more explicit (and more likely to trip you up if you,
e.g., try to `$set` a field to an empty string). This can be overridden
by specifically using the `allowedEmptyString` custom type. And second,
it includes `AllModels`, making it easier for us to write code which
operates on all models (e.g. for future index management tooling).

There are a few smaller regressions in this approach relative to what we
were doing previously. Specifically:
- This will pull in Zod on the client-side. We might be able to fix that
  by pulling mutation methods out of Model, similar to what we're doing
  with `defineMethod`, but it seems annoying and messy so I'm holding
  off right now. Zod isn't that big, anyway, so the footprint impact
  isn't that bad (around 8KB or so).
- There's no way of expressing `denyInsert` or `denyUpdate`. This seems
  fairly difficult to reintroduce, since it's not expressible in JSON
  schema. I'm not sure how useful these were in practice, so I don't
  feel bad about dropping them.
- `Model` does not preserve our `allow` rules which allow admin clients
  to manipulate state. We could re-introduce them, but since those
  default Meteor methods are defined on the collection (and not our
  `Model`), they would completely bypass transformation and thus be
  somewhat unpredictable. I think that instead of supporting these, we
  should build out an admin UI for manipulating state, which would be
  much more predictable and easier to use.

Finally, we update `publishJoinedQuery` to accept either a
`Mongo.Collection` or a `Model`.