Add support for URL strings in schemas
Since our high-level goal is that MongoDB's JSON schema enforcement is
our primary insurance that data in the database matches our expected
schemas, we need to have some enforcement of URL forms in the generated
JSON schema. Ideally, MongoDB would allow us to set `format: "url"` on
the schema, but sadly it does not, so regexes it is.

We tolerate the regex not being perfect for reasons both internal and
external. Externally, it's basically impossible to write a regex that
correctly matches the specification for a valid URL. Internally, it
_should_ be the case that Zod will still validate URLs going into the
database. (Our schema manipulations are vulnerable to fields being
absent that must be present, but should be reasonably robust to fields
being of the wrong type.)