Add a migration to ensure Puzzle tags are unique
Currently all of the cases where we add tags to a puzzle are done using
$addToSet, which should ensure uniqueness, but we have some historical
data (in prod at least) where tags have gotten duplicated.

Historically, we haven't noticed this in practice because we have a
load-time transformation that de-duplicates tags, but our new Zod-based
schema system doesn't support load-time transformations, so we need to
make sure the data in the database is clean. We can do that going
forward by using JSON schema's uniqueItems constraint, but we need to
clean up the old data.