Fix deconflicting of $setOnInsert fields
Mongo will complain if a field in $setOnInsert shows up in any other
part of the update, not just in the $set section. This is a problem due
to how we generate auto-populating fields (including defaults), so be a
little more generous in how we de-conflict potential parallel mutations.

(As an aside, we may need to revisit this for $set fields that are tied
to IsUpdate, but only if we end up trying to, say, $inc an
auto-populating-on-update field; we can punt on that until then.)