Migrate some common test setup into main
For all of our tests, we want rate limits disabled and we want access to
chai-as-promised for promise-related assertions. We might as well set
this up once rather than per-file (and, in fact, disabling rate limits
does cascade out across global state, so we were getting that effect
anyway so long as the smoke tests ran).