Fixup issues discovered by schema validation
When we migrated the database to support multiple answers, we never
actually removed the old (singular) answer property. Additionally, the
Discord API returns "null" for avatar when the user has no avatar, but
we marked the field as optional (meaning it can be absent, but if
present, must be a string).

Add a migration to fix both of those (and logic to handle the avatar
properly going forward).