Experiment with using branded types for hunt IDs
Types in TypeScript are structural by default, meaning that an object of
one type can be assigned to another if those types have the same shape.
Branded types are a form of nominal typing, meaning that an object can't
be assigned to a variable of a branded type unless it is explicitly the
same exact type.

This is useful for types which look the same but should be treated
differently - for instance, a hunt ID and a puzzle ID. They're both
strings, but they're not the same string.

This switches hunt IDs throughout the database to be branded, using
Zod's built-in support for branded types. Zod's branding requires a
brand name; we use the collection name.

This also introduces an assortment of helpers to deal with branded
types:
* `ForeignKeyType<M extends Model>`: The branded type of a foreign key
  to a model. This is used in place of `string` for foreign keys. (Most
  of the changes in this commit are to replace `string` with
  `ForeignKeyType<typeof Hunts>`.)
* `makeForeignKey<M extends Model>(id: string)`: A helper to create a
  foreign key of the correct type. This is used to "encode" a string
  which we know to be a valid foreign key, and should be used very
  sparingly. In practice, we only need this in tests and for the
  hard-coded fixture hunt ID.
* `matchForeignKey<M extends Model>()`: Returns a matcher which can be
  passed to `check` to assert that a value is a foreign key. This is
  actually equivalent to just verifying that the value is any string
  (because we can't actually check that a string is a valid foreign key
  without querying the database), but it's necessary for our handling of
  typed methods and publications.

Note that none of this should introduce any runtime overhead; type
brands exist at the type layer only, and do not have any runtime
footprint (other than values passing through a few extra identity
functions).