Fix find with no arguments
Meteor's find method has the slightly confusing property that passing an
`undefined` query explicitly behaves differently than passing no query
(since the query is an optional parameter). Passing `undefined`
explicitly returns an empty result set.

We want to allow the query to be optional, as with Meteor, but that
resulted in us explicitly passing undefined to Meteor, which caused the
empty result set behavior. So instead make sure we pass an empty query.
(This means that we can no longer access the explicit-undefined behavior
from Model, but I don't think we've ever wanted that anyway.)