Surface when a call transport is in an unexpected state
When clients aren't able to successfully complete the ICE negotiation,
they will transition into the "failed" state. In fact, on the send side,
any state other than "connected" indicates a potential problem (e.g. if
there are no ICE candidates for some reason, the connection will stick
in the "new" state). When this happens, we should display some sort of
error to the user.

This is easy to test by changing the `createWebRtcTransport` call in
`mediasoup.ts` to set both `enableTcp` and `enableUdp` to false
(resulting in no ICE candidates).