Reduce height of metadata section by conditionally showing tag row
Tags are not always the most useful piece of information to show, and
much of the time when working on a puzzle, you'd really rather have a
bit more vertical space for the spreadsheet.

To that end: this patchset makes the tag row of the puzzle metadata
section into a popover, and shows it only when the metadata section
already has the mouse over it or is focused.  This reduces the default
height of the metadata section for an unsolved puzzle from 76 pixels to
48 pixels.

We continue to show correct answers in the metadata section by default
because we consider it important to make it hard to miss that a puzzle
has already been solved and what the answer was, and very few puzzles
involve multiple answers (so the row is not a huge loss).

Fixes #1388.