Unwind Google Drive fetcher if lock is preempted
When we have a lock to fetch Google Drive activity, we attempt to
continually renew it. But if, for some reason, that renewal fails, we
should bail and attempt to reacquire the lock. Previously, we would
blithely go on about our business and continue trying to renew the lock
that we had already lost, resulting in both two processes attempting to
fetch activity history and a lot of error noise from the failed
renewals.