Fix initial firing on Flags.observeChanges
If the flag record does not exist by the time we initialize our
observer, then that means that state is undefined, but the new state
implied by calling checkUpdate would also be undefined, so we'd never
call the callback. We don't actually want to checkUpdate; we want to
call the callback manually.