Use createPortal for modal popups
React's new-ish createPortal allows you to place a component outside of
its rendering tree, which makes sense for modals which are effectively
untethered from where they would appear by default in the DOM.

This is a more stable fix for problems we've had in the past with, e.g.,
overlay triggers, where the modal is inside of the overlay trigger and
thus triggers the tooltip.

For some reason, the structural changes to the components which used
forwardRef caused the react/prop-types eslint rule to start complaining.
Since it's not really useful in a TypeScript world, I just turned it
off.

Fixes #1317.