Add a semi-simple documentation system
Some higher-level information about a codebase are hard to document
inline in code. For systems that span multiple files, functions, etc.,
it's useful to have a more centralized source of documentation. However,
the inherent risk of any such centralized documentation is that it
quickly becomes out of date.

Introduce a `docs` folder, containing such higher-level documentation.
However, to mitigate against docs becoming out of date, additionally
introduce some validation in the form of a new, homegrown linter. This
linter requires that some metadata be included with each documentation
file: specifically, which files the documentation covers and when the
documentation was last updated. If the files are ever updated without a
corresponding update to the docs, the linter will fail.