Coalesce dependabot updates
Dependabot has support for grouping updates based on various
characteristics. Given that most package updates either require no code
changes or are backwards-incompatible and should be ignored, we can
reduce PR churn by opening a single PR with all pending changes.

In order to make this more effective, also push updates to happen
weekly, instead of daily. I picked Fridays, as that tends to be when I'm
most available to look at PRs, and leads into the weekend for maximum
flexibility.

Note that none of this applies to security updates, which dependabot
will still open immediately.