Configure dependabot ignores by file
This both gives us better persistent tracking of updates that we
intentionally aren't taking and also gives us a workaround for a current
bug where the "@dependabot ignore" directive doesn't work for
"@"-prefixed packages.

Once merged, I'll go through and issue unignore directives for each of
these, so that they don't override the ignore directives in the config
file.