Pull Google OAuth2 client direct from dependencies
Historically, we've depended on google-auth-library directly in order to
get our hands on an OAuth2 client, but this has been a bit of a pain, as
the Google API modules have not always been kept in sync with the latest
version of google-auth-library, causing incompatibilities across version
upgrades.

However, each of the modules exports an auth object which includes the
OAuth2 client, so we can just pick one of the libraries (I picked drive,
since that's our most heavily used API) and use that instead.