Allow declaring methods and publications under models
Almost all of our publications and models are associated (or at least
primarily associated) with a model. This is reflected in the way we name
our methods ("verbNoun"). However, we previously placed all methods in a
single flat directory and did the same with publications.

Instead, allow declaring methods and publications as part of creating a
new model class. This has the convenient side-effect of reducing the
imports you need to operate on a model to just the model class itself.
It also makes the internal identifier for a method or publication match
how we refer to it in code (e.g. `ChatMessages.methods.send`).

For now, the implementations of the methods and publications are still
in the same files as before; that's potentially something to reevaluate
in the future. Additionally, it arguably makes it harder to enumerate
all methods and publications, but that might be an acceptable tradeoff.