Add type declaration for styled react-bootstrap Buttons
This works around a bug in either styled-components or react-bootstrap
(slightly unclear which) that causes an error ("Expression produces a
union type that is too complex to represent.") with the latest versions
of both.

It's not entirely clear to me why this only seems to be required for
Buttons and not other react-bootstrap components, but this seems to make
tsc happy.