Take all updates that we can
The latest versions of mediasoup use the `??=` operator, which isn't
supported on Node 14, so we need to block dependabot updates.
Additionally, our block on `@types/node` didn't allow us to take updates
to the v14 types, so relax that slightly.

Additionally, drop `p-throttle` as a dependency, as we have not used it
since c2d07b3485e21e329926d8c0dbd89db4caa20990.

Enable `skipLibCheck` on our eslint plugin - this causes tsc to
typecheck our plugin code but take library types as accurate. This is
necessary because our version of typescript-eslint has some
incompatibilities with the latest version of TypeScript, and we can't
currently update typescript-eslint due to Node v14 incompatibilities.

Finally, some update-related housekeeping: add aria-labels to components
in `GuessQueuePage` and `Puzzle` which have tooltips (which aren't
typically used as labels by screenreaders). These aren't required with
our current version of `eslint-plugin-jsx-a11y`, but it seems like they
will likely be required in the future. (The latest verison does require
them, although it also has a bug with table elements, so we are
deferring that update for now.)