Make no-disallowed-sync-methods stricter
This now disallows some additional sync methods. As far as I can tell,
this should include all sync methods that currently have an async
alternative.

I also had to mess around with how TypedMethod is declared, as the
method we're using to get the fully-qualified type name behaves oddly on
types which are exported in the same statement as their declaration.