Prepare to switch to prettier for code formatting
This adds prettier as a dependency and lint step (and recommended VSCode
extension), reconfigures eslint with the prettier config package (so
that it stops linting for formatting that is managed by prettier), and
drops a handful of eslint disables that are no longer needed (since
prettier will supercede them).

And finally - in a bit of inside baseball - it does some preparatory
formatting of the `tests/unit/imports/server/Model.ts` file. Once
prettier reformats this file, it would move the object members onto the
next line, thus moving them _out_ of line with the @ts-ignore comment.
By pre-splitting them now, we keep them in sync.