Fix rendering of post-username migration Discord handles
It looks like we still see a mix of Discord account references that both
have and lack discriminators. At some point in the future once all
usernames have been migrated, we may no longer need to track the
discriminator at all, but for the time being only display it if it
represents a non-migrated value (i.e. is not "0").