Allow track IDs to be opaque strings
We assumed these would always be UUIDs, but it seems like at least
modern Firefox returns track IDs that are not (pure) UUIDs. (They seem
to have curly braces around them).

Since we just use the track ID to connect the ProducerClient and
ProducerServer records, we need them to be present, but we can treat
them as opaque.