Move JR nav items into top nav bar
I can never ever see the blue items menu even when looking for it--
partly because it's not in an expected place for navigation,
partly because it gets stretched super wide on non-small windows such
that it doesn't even look like a set of buttons.

Now they're in the nav bar (if a hunt ID is set; they don't appear
on the hunts index page).

Also improves contrast on the top nav bg a little and adds a bottom border
to the nav bar (in the same color as the vertical panes separator).

In small windows the menu appears in the old spot and  looks more or
less like the old one.